CSP-2706 Changed to SystemJsonSerializer and update event model

+++ src/SFA.DAS.RoATPService.Application/Events/ProviderStatusChangedEvent.cs
@@ -2,6 +2,6 @@ namespace SFA.DAS.RoATPService.Application.Events;
 
 public class ProviderStatusChangedEvent
 {
-    public long Ukprn { get; set; }
+    public int Ukprn { get; set; }
     public string Status { get; set; } = string.Empty;
 }
